+++ changelog.md
@@ -1,2 +1,8 @@
+## v1.0.1
+- **Added** trigger whitelist to settings
+- **Added** named editor groups disclaimer
+
+- **Fixed** layer alpha not even being implemented
+
 ## v1.0.0
 - Initial Release